Commit of changes to propagators and integrators code; addition of unit tests; minor changes to other code.
git-svn-id: file:///srv/svn/repos/tudat/trunk@93 ac7aa084-b6da-11df-bf42-7fac39f0224f

+++ Mathematics/NumericalIntegrators/integrator.cpp
@@ -276,8 +276,9 @@ void Integrator::computeInternalDerivedIntegrationParameters_( )
                                - integrationIntervalStart_;
 
         // Compute number of integration steps for fixed stepsize methods.
-        numberOfIntegrationSteps_ = std::ceil( integrationInterval_
-                                               / stepsize_ );
+        numberOfIntegrationSteps_ = static_cast < unsigned int >
+                                    ( std::ceil( integrationInterval_
+                                                 / stepsize_ ) );
 
         // Set dimension of initial state.
         dimensionOfCurrentState_ = vectorOfCurrentStatePointers_.at( 0 )
